requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-registration==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.12.1
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-mimeparse==1.6.0
pytz==2020.4
regex==2020.10.28
requests==2.24.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.2.0
